+++ packages/flame/test/game/base_game_test.dart
@@ -8,7 +8,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart' as flutter;
-import 'package:test/test.dart';
+import 'package:flutter_test/flutter_test.dart';
 
 class MyGame extends FlameGame with HasTappableComponents {}
 
@@ -222,5 +222,23 @@ void main() {
     expect(game.children.length, equals(3));
     game.update(0.0);
     expect(game.children.isEmpty, equals(true));
+  });
+
+  test("can't add a component to a game that don't have layout yet", () {
+    final game = MyGame();
+    final component = MyComponent();
+
+    const message = '"prepare/add" called before the game is ready. '
+        'Did you try to access it on the Game constructor? '
+        'Use the "onLoad" ot "onParentMethod" method instead.';
+
+    expect(
+      () => game.add(component),
+      throwsA(
+        predicate(
+          (e) => e is AssertionError && e.message == message,
+        ),
+      ),
+    );
   });
 }