+++ packages/flame/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Next]
  - `changePriority` no longer breaks game loop iteration
+ - Move component mixin checks to their own files
 
 ## [1.0.0-releasecandidate.15]
  - Fix issue with `Draggable`s not being removed from `draggables` list