+++ packages/flame/CHANGELOG.md
@@ -5,7 +5,7 @@
  - Move component mixin checks to their own files
  - Fix exception when game was rebuilt
  - Add `@mustCallSuper` on `Component.render`
- - Add `SpriteSheet.createAnimationVariable` method to allow animations with different stepTime for each sprite
+ - Add `SpriteSheet.createAnimationVariable` method to allow animations with different `stepTime` for each sprite
 
 ## [1.0.0-releasecandidate.15]
  - Fix issue with `Draggable`s not being removed from `draggables` list