+++ packages/flame/CHANGELOG.md
@@ -5,6 +5,7 @@
  - Move component mixin checks to their own files
  - Fix exception when game was rebuilt
  - Add `@mustCallSuper` on `Component.render`
+ - Use the full delta in `JoystickComponent` so that it can't go to the wrong direction on the wrong side
 
 ## [1.0.0-releasecandidate.15]
  - Fix issue with `Draggable`s not being removed from `draggables` list